+++ packages/pluggableWidgets/video-player-web/src/components/Html5.tsx
@@ -1,5 +1,7 @@
 import { Component, createElement, createRef } from "react";
 
+import classNames from "classnames";
+
 export interface Html5PlayerProps {
     url: string;
     poster?: string;
@@ -9,6 +11,7 @@ export interface Html5PlayerProps {
     muted: boolean;
     style?: any;
     aspectRatio?: boolean;
+    preview: boolean;
 }
 
 export class Html5 extends Component<Html5PlayerProps> {
@@ -19,10 +22,31 @@ export class Html5 extends Component<Html5PlayerProps> {
 
     render(): JSX.Element {
         return (
-            <div className="widget-video-player-html5-container">
-                <div className="video-error-label-html5" ref={this.errorElement}>
-                    The video failed to load :(
-                </div>
+            <div
+                className={classNames("widget-video-player-html5-container", {
+                    "widget-video-player-show-controls": this.props.showControls
+                })}
+            >
+                {this.props.preview ? (
+                    <svg
+                        className="widget-video-player-preview-play-button"
+                        width="48"
+                        height="48"
+                        viewBox="0 0 48 48"
+                        fill="none"
+                        xmlns="http://www.w3.org/2000/svg"
+                    >
+                        <path
+                            d="M0 24C0 10.7452 10.7452 0 24 0C37.2548 0 48 10.7452 48 24C48 37.2548 37.2548 48 24 48C10.7452 48 0 37.2548 0 24Z"
+                            fill="#373737"
+                        />
+                        <path d="M16 12V36L34.8571 24L16 12Z" fill="white" />
+                    </svg>
+                ) : (
+                    <div className="video-error-label-html5" ref={this.errorElement}>
+                        The video failed to load :(
+                    </div>
+                )}
                 <video
                     className="widget-video-player-html5"
                     controls={this.props.showControls}
@@ -34,12 +58,14 @@ export class Html5 extends Component<Html5PlayerProps> {
                     height={!this.props.aspectRatio ? "100%" : undefined}
                     preload={this.props.poster ? "metadata" : "auto"}
                 >
-                    <source
-                        src={this.props.url}
-                        type="video/mp4"
-                        onError={this.handleOnError}
-                        onLoad={this.handleOnSuccess}
-                    />
+                    {!this.props.preview ? (
+                        <source
+                            src={this.props.url}
+                            type="video/mp4"
+                            onError={this.handleOnError}
+                            onLoad={this.handleOnSuccess}
+                        />
+                    ) : null}
                 </video>
             </div>
         );