+++ packages/pluggableWidgets/video-player-web/CHANGELOG.md
@@ -13,4 +13,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   We renamed the video URL and Poster URL widget properties.
 -   We improved property rendering in Studio.
--   We removed rendering errors inside design mode preview.
+-   We improved the design mode preview.