+++ packages/tools/pluggable-widgets-tools/CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this tool will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 9.3.0 - 2021-07-02
+### Added
+- We added support for testing components that make use of SVG images.
 
 ## 9.2.1 - 2021-06-16
 - We reintroduced the possibility to use `.env` file for environment variables (removed in v9 previously)